feat: implement modular refactoring, block-level checkpointing, and self-contained runs flow
