--- HEARTBEAT.md
@@ -1,9 +1,0 @@
-```markdown
-# Keep this file empty (or with only comments) to skip heartbeat API calls.
-
-# Add tasks below when you want the agent to check something periodically.
-```
-
-## Related
-
-- [Heartbeat config](/gateway/config-agents)
