--- training_run_v1.csv
@@ -1,30 +1,0 @@
-step,loss,ce,avg_ponder,avg_forget_cost,t_total,data_wait,compute_time,logits_mean,logits_std,logits_min,logits_max,prob_mean,prob_std,saturation,temporal_drift,forget_density,logit_spread,diversity_loss
-100,2.2567,8.5350,1.06,0.4919,0.42,0.0010,0.4138,2.7944,0.9051,0.2140,5.1083,9.0644,0.4015,0.9995,0.8685,0.4922,4.8943,0.9833
-200,2.0652,7.7692,21.03,0.4788,0.42,0.0008,0.4137,-6.0621,3.1502,-13.1299,-2.4251,29.0275,0.0127,0.9993,0.6688,0.4785,10.7048,0.9830
-300,1.8615,7.3683,23.98,0.4934,0.42,0.0010,0.4137,-9.8248,1.3209,-12.5569,-5.6124,31.9789,0.0004,0.2723,1.2770,0.4922,6.9446,0.1556
-400,1.6823,6.7090,23.95,0.4961,0.42,0.0009,0.4135,-8.5582,0.9205,-10.5515,-5.2879,31.9494,0.0005,0.1252,1.1590,0.4980,5.2636,0.0404
-500,1.5208,6.0716,23.96,0.4980,0.42,0.0009,0.4137,-8.8696,1.1457,-11.5251,-4.3731,31.9600,0.0009,0.0888,1.1942,0.4980,7.1520,0.0233
-600,1.5169,6.0584,23.91,0.4964,0.42,0.0008,0.4136,-8.1986,1.3917,-12.2996,-2.9704,31.9075,0.0034,0.0880,1.2788,0.4961,9.3291,0.0185
-700,1.4735,5.8877,23.83,0.4993,0.42,0.0008,0.4137,-7.9675,1.4697,-11.2440,-0.9285,31.8319,0.0102,0.0753,1.1943,0.4980,10.3155,0.0127
-800,1.3246,5.2931,23.99,0.5000,0.42,0.0008,0.4134,-10.3415,1.3895,-12.9472,-4.5886,31.9933,0.0008,0.0719,1.2241,0.4980,8.3587,0.0104
-900,1.2940,5.1720,23.99,0.5016,0.42,0.0008,0.4133,-9.8567,1.3129,-12.6783,-4.0965,31.9881,0.0013,0.0633,1.2202,0.5000,8.5818,0.0082
-1000,1.3277,5.3073,23.98,0.5015,0.42,0.0008,0.4137,-9.2391,1.4242,-12.6558,-3.0611,31.9754,0.0034,0.0626,1.2505,0.4980,9.5947,0.0071
-1100,1.2731,5.0893,23.97,0.5012,0.42,0.0007,0.4134,-9.0123,1.5095,-12.4579,-3.2509,31.9681,0.0047,0.0588,1.2692,0.4980,9.2070,0.0060
-1200,1.1000,4.3975,23.99,0.5015,0.42,0.0007,0.4139,-10.4502,1.6692,-13.8645,-3.3252,31.9926,0.0027,0.0545,1.2778,0.4980,10.5393,0.0053
-1300,1.3388,5.3524,23.99,0.5027,0.42,0.0008,0.4141,-9.9915,1.5586,-12.5912,-3.0294,31.9898,0.0038,0.0567,1.3117,0.4980,9.5618,0.0052
-1400,1.2102,4.8380,23.98,0.5022,0.42,0.0008,0.4138,-9.4733,1.6038,-13.1330,-2.6831,31.9811,0.0052,0.0581,1.3506,0.4980,10.4499,0.0052
-1500,1.1251,4.4984,23.99,0.5025,0.42,0.0009,0.4140,-10.0152,1.5988,-12.6706,-2.9931,31.9897,0.0044,0.0531,1.3849,0.4980,9.6776,0.0042
-1600,1.1048,4.4167,23.99,0.5043,0.42,0.0011,0.4154,-10.2398,1.7724,-13.7530,-2.5990,31.9903,0.0051,0.0589,1.3688,0.4980,11.1540,0.0054
-1700,1.2264,4.9036,23.98,0.5008,0.42,0.0007,0.4137,-9.4934,1.6004,-12.2973,-2.6870,31.9824,0.0072,0.0566,1.3614,0.4961,9.6103,0.0043
-1800,1.1605,4.6402,23.99,0.5030,0.42,0.0008,0.4136,-10.1425,1.6741,-13.1606,-2.9586,31.9912,0.0052,0.0488,1.3800,0.4980,10.2020,0.0035
-1900,1.3087,5.2330,23.99,0.5025,0.42,0.0008,0.4136,-9.9947,1.7375,-12.6323,-2.1280,31.9883,0.0075,0.0527,1.3707,0.4961,10.5044,0.0040
-2000,1.3205,5.2799,23.99,0.5017,0.42,0.0007,0.4136,-9.8062,1.6513,-12.3978,-2.4098,31.9874,0.0074,0.0539,1.3501,0.4961,9.9880,0.0039
-2100,1.2122,4.8470,23.98,0.5034,0.42,0.0008,0.4138,-9.6245,1.7473,-12.9759,-1.9046,31.9840,0.0113,0.0499,1.3558,0.4961,11.0713,0.0036
-2200,1.5395,6.1562,23.98,0.5039,0.25,0.0005,0.2444,-9.2994,1.6267,-11.4926,-2.0662,31.9800,0.0108,0.0502,1.3646,0.4980,9.4264,0.0034
-2300,1.5471,6.1865,23.98,0.5039,0.25,0.0005,0.2444,-9.4592,1.6916,-12.2849,-1.4223,31.9810,0.0125,0.0543,1.3775,0.4980,10.8626,0.0039
-2400,1.4318,5.7254,23.99,0.5036,0.25,0.0005,0.2447,-9.6379,1.6240,-11.8724,-2.6945,31.9855,0.0068,0.0511,1.3732,0.4980,9.1779,0.0035
-2500,1.5095,6.0362,23.98,0.5039,0.25,0.0005,0.2451,-9.3349,1.5992,-11.7362,-2.1172,31.9811,0.0090,0.0517,1.3722,0.4980,9.6190,0.0036
-2600,1.4786,5.9126,23.98,0.5029,0.25,0.0005,0.2450,-9.3295,1.5918,-11.5102,-2.6007,31.9812,0.0090,0.0493,1.3735,0.4961,8.9094,0.0033
-2700,1.4865,5.9442,23.98,0.5034,0.25,0.0005,0.2445,-9.2985,1.6074,-11.6255,-2.5187,31.9806,0.0093,0.0498,1.3671,0.4980,9.1067,0.0034
-2800,1.4761,5.9027,23.98,0.5039,0.25,0.0005,0.2446,-9.5333,1.6379,-11.8997,-2.3991,31.9844,0.0084,0.0501,1.3688,0.4980,9.5006,0.0034
-2900,1.4753,5.8994,23.98,0.5036,0.25,0.0005,0.2442,-9.4233,1.5944,-11.6443,-2.4263,31.9831,0.0079,0.0500,1.3756,0.4980,9.2180,0.0034
